# backend-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/users/me": {
            "get": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Get current user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get current user profile",
                "responses": {
                    "200": {
                        "description": "Current user",
                        "schema": {
                            "$ref": "#/definitions/user.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Update current user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update current user profile",
                "parameters": [
                    {
                        "description": "User update payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated user",
                        "schema": {
                            "$ref": "#/definitions/user.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "Register a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created user and JWT token",
                        "schema": {
                            "$ref": "#/definitions/user.RegisterUserResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "response.ErrorResponse": {
            "description": "Error response structure",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "user.RegisterUserRequest": {
            "description": "User registration payload",
            "type": "object",
            "required": [
                "email",
                "id"
            ],
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "plan": {
                    "type": "string"
                }
            }
        },
        "user.RegisterUserResponse": {
            "description": "Created user and JWT token",
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJSUzI1NiIs..."
                },
                "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJSUzI1NiIs..."
                },
                "user": {
                    "$ref": "#/definitions/user.User"
                }
            }
        },
        "user.RegisterUserResponseWrapper": {
            "description": "Registration response wrapped in response envelope",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/user.RegisterUserResponse"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "user.TeamInfo": {
            "description": "Team information",
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "canCreateProject": {
                    "type": "boolean"
                },
                "memberCount": {
                    "type": "integer"
                },
                "personal": {
                    "type": "boolean"
                },
                "plan": {
                    "type": "string"
                },
                "projectCount": {
                    "type": "integer"
                },
                "teamId": {
                    "type": "string"
                },
                "teamName": {
                    "type": "string"
                }
            }
        },
        "user.UpdateUserRequest": {
            "description": "User update payload",
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "plan": {
                    "type": "string"
                }
            }
        },
        "user.User": {
            "description": "User profile",
            "type": "object",
            "properties": {
                "activeTeamId": {
                    "type": "string"
                },
                "avatarUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "plan": {
                    "type": "string"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.TeamInfo"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "user.UserResponse": {
            "description": "User wrapped in response envelope",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/user.User"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    }
}